requirements.txt
django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
openai==0.27.0
psycopg2
drf-spectacular[sidecar]
pytest-django
pytest-cov
pytest-mock
django-jazzmin
django-celery-beat==2.4.0
celery==5.2.7
redis==4.3.4
flower==1.2.0
Pillow
